Incorporated last changes from issue-154 branch
