+++ src/SharpBrick.PoweredUp.Mobile.Examples.iOS/SharpBrick.PoweredUp.Mobile.Examples.iOS.csproj
@@ -67,10 +67,13 @@
     <LangVersion>latest</LangVersion>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="ApplicationWrapper.cs" />
+    <Compile Include="IosInitializer.cs" />
     <Compile Include="Main.cs" />
     <Compile Include="AppDelegate.cs" />
     <None Include="Entitlements.plist" />
     <None Include="Info.plist" />
+    <Compile Include="NativeDeviceInfoProvider.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
